RUB-263 rust: keep OpenSSL policy facade
